Added example hamcrest

+++ src/main/java/Calculator.java
@@ -17,7 +17,7 @@ public class Calculator {
     //    }
     //};
 
-    BinaryOperator<Integer> devide = (x, y) -> x / y;
+    BinaryOperator<Float> devide = (x, y) -> x / y;
     UnaryOperator<Integer> pow = x -> x * x;
     UnaryOperator<Integer> abs = x -> x > 0 ? x : x * -1;
     Predicate<Integer> isPositive = x -> x > 0;
